requirements.txt
absl-py @ file:///tmp/build/80754af9/absl-py_1607439979954/work
aiohttp @ file:///tmp/build/80754af9/aiohttp_1602530275854/work
appdirs==1.4.4
astunparse==1.6.3
async-timeout==3.0.1
attrs @ file:///tmp/build/80754af9/attrs_1604765588209/work
backcall==0.2.0
bdlb @ git+https://github.com/hermannsblum/bdl-benchmark.git@099c95eb75897b3cc9be74484f11357c0989e3a7
blinker==1.4
brotlipy==0.7.0
cachetools @ file:///tmp/build/80754af9/cachetools_1611600262290/work
certifi==2020.12.5
cffi @ file:///tmp/build/80754af9/cffi_1613246439577/work
chardet==4.0.0
cityscapesScripts==2.1.7
click @ file:///home/linux1/recipes/ci/click_1610990599742/work
cloudpickle==1.6.0
colorama==0.4.4
coloredlogs==15.0
coverage @ file:///tmp/build/80754af9/coverage_1611690792189/work
cryptography @ file:///tmp/build/80754af9/cryptography_1607635303852/work
cycler==0.10.0
Cython==0.29.21
dataclasses==0.8
decorator==4.4.2
detectron2 @ git+https://github.com/facebookresearch/detectron2.git@45a8bfb64053d71d9d7f136fb25a6abe841dc91f
dill==0.3.3
dm-tree==0.1.5
docopt==0.6.2
future==0.18.2
fvcore==0.1.2.post20210128
gast==0.4.0
gitdb==4.0.5
GitPython==3.1.11
google-auth @ file:///tmp/build/80754af9/google-auth_1607969906642/work
google-auth-oauthlib @ file:///tmp/build/80754af9/google-auth-oauthlib_1603929124518/work
google-pasta==0.2.0
googleapis-common-protos==1.52.0
grpcio @ file:///home/builder/ktietz/ci_grpcio/grpcio_1613123281713/work
h5py @ file:///tmp/build/80754af9/h5py_1593454121459/work
hdbscan==0.8.26
humanfriendly==9.1
idna @ file:///home/linux1/recipes/ci/idna_1610986105248/work
idna-ssl @ file:///tmp/build/80754af9/idna_ssl_1611752490495/work
imageio==2.9.0
importlib-metadata==3.3.0
importlib-resources==5.1.0
iopath==0.1.3
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.18.0
joblib==1.0.0
jsonpatch==1.28
jsonpickle==1.4.2
jsonpointer==2.0
kaggle==1.5.10
Keras-Preprocessing @ file:///tmp/build/80754af9/keras-preprocessing_1612283640596/work
kiwisolver==1.3.1
libmr==0.1.9
Markdown @ file:///tmp/build/80754af9/markdown_1605111083232/work
matplotlib==3.3.3
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
multidict @ file:///tmp/build/80754af9/multidict_1600456401639/work
munch==2.5.0
natsort==7.1.1
networkx==2.5
numpy==1.19.5
oauthlib==3.1.0
opencv-python==4.4.0.46
opt-einsum==3.1.0
packaging==20.8
pandas==1.1.5
panopticapi @ git+https://github.com/cocodataset/panopticapi.git@118ffaba478b5d45925e88d504625b4123f3160a
parso==0.8.1
patsy==0.5.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
portalocker==2.2.1
promise==2.3
prompt-toolkit==3.0.16
protobuf==3.14.0
psutil==5.8.0
ptyprocess==0.7.0
py-cpuinfo==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pydensecrf==1.0rc3
pydot==1.4.1
Pygments==2.8.0
PyJWT==1.7.1
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1608057966937/work
pyparsing==2.4.7
pyquaternion==0.9.9
PySocks @ file:///tmp/build/80754af9/pysocks_1605305763431/work
python-dateutil==2.8.1
python-slugify==4.0.1
pytz==2020.5
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==21.0.1
requests @ file:///tmp/build/80754af9/requests_1608241421344/work
requests-oauthlib==1.3.0
rsa @ file:///tmp/build/80754af9/rsa_1610483308194/work
sacred==0.8.2
scikit-image==0.17.2
scikit-learn==0.24.1
scipy==1.2.1
seaborn==0.11.1
six==1.15.0
sklearn==0.0
smmap==3.0.4
statsmodels==0.12.1
tabulate==0.8.7
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.1
tensorflow==2.2.0
tensorflow-datasets==3.0.0
tensorflow-estimator==2.2.0
tensorflow-metadata==0.27.0
termcolor==1.1.0
text-unidecode==1.3
tfp-nightly==0.13.0.dev20210208
threadpoolctl==2.1.0
tifffile==2020.9.3
torch @ file:///data1/users/caijh28/download/torch-1.5.0-cp36-cp36m-linux_x86_64.whl
torchfile==0.1.0
torchvision==0.6.0
tornado==6.1
tqdm==4.55.0
traitlets==4.3.3
typing==3.7.4.3
typing-extensions @ file:///home/ktietz/src/ci_mi/typing_extensions_1612808209620/work
urllib3 @ file:///tmp/build/80754af9/urllib3_1611694770489/work
visdom==0.1.8.9
wcwidth==0.2.5
websocket-client==0.57.0
Werkzeug @ file:///home/ktietz/src/ci/werkzeug_1611932622770/work
wrapt==1.12.1
yacs==0.1.8
yarl @ file:///tmp/build/80754af9/yarl_1606939915466/work
zipp @ file:///tmp/build/80754af9/zipp_1604001098328/work
